src/auth_service/auth_service/requirements.txt
fastapi[all]==0.95.0
pydantic==1.10.6
pydantic[email]
SQLAlchemy==2.0.7
alembic==1.10.2
uvicorn==0.21.1
gunicorn
passlib[bcrypt]
tenacity==8.2.2
python-multipart==0.0.5
asyncpg==0.27.0
asyncpg_utils==0.6.0
psycopg2-binary
psycopg2
python-jose==3.3.0
pytest==7.2.0
bcrypt==4.0.1
httpx==0.23.0
python-dotenv
loguru==0.6.0
bson
black==23.1.0
faker==18.3.0
pytest-asyncio
pytest-cov
pre-commit
fastapi-mail
asgi-correlation-id
locust
